requirements.txt
django-tastypie==0.12.2
